+++ libs/cailie/output.h
@@ -13,7 +13,7 @@ class CaOutput {
 		void child(const std::string &name);
 		void back();
 		void set(const std::string &name, const int i);
-		void set(const std::string &name, const unsigned int i);
+		void set(const std::string &name, const size_t i);
 		void set(const std::string &name, const std::string &s);
 		void set(const std::string &name, const char *s) {
 			set(name, std::string(s));
